AfriMCQA/pyproject.toml
[project]
name = "afrimcqa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "notebook>=7.5.5",
    "ipywidgets>=8.0.0",
    # "numpy<2.0",
    # "pybind11>=2.12",
    # "torch==2.2.2",
    "torch>=2.10.0",
    "transformers>=5.3.0",
    "pillow>=12.1.1",
    "torchvision>=0.25.0",
    "openai>=2.30.0",
    "pydantic>=2.12.5",
    "dotenv>=0.9.9",
    "datasets>=4.8.4",
    "torchcodec>=0.11.0",
    "google-cloud-storage>=2.0",
    "qwen-vl-utils>=0.0.14",
    "peft>=0.18.1",
    "requests>=2.32.5",
    "accelerate>=1.13.0",
]
